requirements.txt
Werkzeug>=0.9
click==4.0
python-easyconfig==0.1.0
